+++ CQueue.cpp
@@ -21,8 +21,8 @@ CQueue::CQueue()
     //
     // this is required for maintaining order of the messages
     
-    srand(time(NULL));
-    long queue_id = rand();
+    srand((uint)time(NULL));
+    uint queue_id = rand();
     queue_ref = msgget(queue_id,  IPC_CREAT | 0666);
     queue_cont_ref = msgget(1111, IPC_CREAT | 0666);
     
@@ -75,7 +75,7 @@ bool CQueue::send_msg(string msg, bool use_p_type, long type) const
     //calculate random type number for following parts
     //
     //required for identifying message
-    srand(time(NULL));
+    srand((uint)time(NULL));
     int cont_type = rand() %8000 +1;
